# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  financetracker_internal_application_dto.BudgetListResponse:
    properties:
      budgets:
        items:
          $ref: '#/definitions/financetracker_internal_application_dto.BudgetResponse'
        type: array
      total_count:
        type: integer
    type: object
  financetracker_internal_application_dto.BudgetResponse:
    properties:
      amount:
        type: number
      category_id:
        type: string
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      period:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  financetracker_internal_application_dto.CreateBudgetRequest:
    properties:
      amount:
        type: number
      category_id:
        type: string
      end_date:
        type: string
      period:
        enum:
        - monthly
        - yearly
        type: string
      start_date:
        type: string
    required:
    - amount
    - category_id
    - period
    - start_date
    type: object
...
  financetracker_internal_application_dto.UpdateBudgetRequest:
    properties:
      amount:
        type: number
      end_date:
        type: string
      is_active:
        type: boolean
      start_date:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@financetracker.local
    name: API Support
  description: Finance management REST API with Auth0 authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: FinanceTracker API
  version: "1.0"
paths:
...
  /api/v1/budgets:
    get:
      consumes:
      - application/json
      description: ユーザーの予算一覧を取得します
      parameters:
      - description: カテゴリID
        in: query
        name: category_id
        type: string
      - description: 期間
        enum:
        - monthly
        - yearly
        in: query
        name: period
        type: string
      - description: アクティブ状態
        in: query
        name: is_active
        type: boolean
      - default: '"start_date desc"'
        description: ソート順
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/financetracker_internal_application_dto.BudgetListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 予算一覧取得
      tags:
      - budgets
    post:
      consumes:
      - application/json
      description: 新しい予算を作成します
      parameters:
      - description: 作成する予算情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/financetracker_internal_application_dto.CreateBudgetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/financetracker_internal_application_dto.BudgetResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 予算作成
      tags:
      - budgets
  /api/v1/budgets/{id}:
    delete:
      consumes:
      - application/json
      description: 指定されたIDの予算を削除します
      parameters:
      - description: 予算ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 予算削除
      tags:
      - budgets
    get:
      consumes:
      - application/json
      description: 指定されたIDの予算を取得します
      parameters:
      - description: 予算ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/financetracker_internal_application_dto.BudgetResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 予算取得
      tags:
      - budgets
    put:
      consumes:
      - application/json
      description: 指定されたIDの予算を更新します
      parameters:
      - description: 予算ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新する予算情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/financetracker_internal_application_dto.UpdateBudgetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/financetracker_internal_application_dto.BudgetResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 予算更新
      tags:
      - budgets
  /api/v1/budgets/current:
    get:
      consumes:
      - application/json
      description: 現在の日付に対してアクティブな予算を取得します
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/financetracker_internal_application_dto.BudgetListResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 現在の期間の予算取得
      tags:
      - budgets
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"